Unfinished attempt to use lua integers
Problem: apparently there is no C function in luajit that will produce luajit 
integers (those represented by 10LL and 10ULL). Worse, there is a special type 
for them, but not type constant and no function that does something useful with 
them. This problem should be deferred to the time when converter.c and 
msgpack_lua_c_bindings.c (and, maybe, something else) will be rewritten using 
luajit FFI.

Code was not tested with lua 5.3, though it pretends it was.