parser/printer/tests: Add various commands parsing support
Added:

- `:autocmd`
- `:*unmenu`
- `:buffer` and some other `:b*` buffer commands
- `:behave`
- `:breakadd`/`:breakdel`/`:profile`
- `:cbuffer` and friends
- `:left`/`:right`/`:center`
- `:checktime`
- `:clist`/`:llist`
- `:ilist` and friends
- `:catch`
- `:copy`/`:move`/`:t`
- `:command`
- `:yank`/`:delete`/`:put`: REGSTR commands
- `:!`
- `:debug`
- `:delcommand`
- `:display`/`:registers`
- `:diffget`/`:diffput`
- `:digraphs`
- `:doautocmd`
- `:doautoall`
- `:drop`
- `:earlier`/`:later`
- `:filetype`
- `:history`
- `:mark`/`:k`
- `:*make`/`:*grep*`
- `:nbstart`
- `:popup`
- `:retab`
- `:resize`
- `:redir`
- various script commands
- `:mode`
- `:open`
- `:profdel`
- `:*select` and `:*jump`
- `:[v]global`
- `:*vimgrep*`
- `:marks`
- `:match`
- `:py3`
- `:sbuffer`
- `:set*`
- `:sleep`
- `:substitute` and friends
- `:sort`
- `:spell*`
- `:syntime`
- `:tabmove` (with function for `:resize`, slightly modified)
- `:tearoff` (with function for `:popup`)
- `:winsize`/`:winpos`
- `:wincmd`
- `:z`
- `:*`/`:@`
- `:&`/`:~`
- `:wsverb`
- `:help`
- `:*helpgrep`

Note: kArgAddress is now using Range *range union member, but is not named
kArgRange. Reason: it still represents *one* address, but I have to use Range
structure due to `10;20` being a valid address (will use line 20, but in case of
`10;/abc` first part should be taken into account: this variant will use line
that matches `/abc` that is located after line `10`).

Note 2: last :command argument is not parsed. Reasoning: things like <args> need
to be expanded before parsing anything. I can redirect parsing of some things
like <f-args> to expressions parser or <line1> to range parser, but this is not
compatible and is not going to work for things like

    command -nargs=1 Sil silent! <args>