printer: Make it possible to override written newlines
Reasoning: to proceed with first stage of #1536 it is needed to supply old 
parser correct `fgetline` function. But it may appear that dumped newline is 
actually part of the command line (e.g. when parsing `:execute "normal 
iabc\ndef"`). To distinguish newlines between commands with in-command ones 
ability to override in-command newline with some string that contains NUL is 
needed (if string does *not* contain NUL it is still possible for it to appear 
inside `normal` command).